requirements.txt
flask
flask-cors
pystray
Pillow
pyinstaller
cryptography
pynfe
lxml
signxml
pyOpenSSL
requests
tzdata
python-escpos

pywin32; sys_platform == 'win32'
pycups; sys_platform == 'linux'